# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/orders": {
      "get": {
        "operationId": "AdminOrdersController_list",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "DRAFT",
                "PENDING_PAYMENT",
                "PAID",
                "CONFIRMED",
                "PARTIALLY_FULFILLED",
                "FULFILLED",
                "CANCELLED",
                "REFUNDED"
              ]
            }
          },
          {
            "name": "paymentMethod",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "online",
                "cod"
              ]
            }
          },
          {
            "name": "shopId",
            "required": false,
            "in": "query",
            "schema": {
              "example": "15",
              "type": "string"
            }
          },
          {
            "name": "dateFrom",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2026-08-01",
              "type": "string"
            }
          },
          {
            "name": "dateTo",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2026-08-31",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "example": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 20,
              "example": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{ items, total, page, limit, totalPages }"
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Barcha buyurtmalar (status/to‘lov/do‘kon/sana filtr + sahifalash)",
        "tags": [
          "admin-orders"
        ]
      }
    },
    "/api/v1/admin/orders/{id}": {
      "get": {
        "operationId": "AdminOrdersController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Buyurtma to‘liq (sub-order + item + shipment + to‘lov)",
        "tags": [
          "admin-orders"
        ]
      }
    },
    "/api/v1/admin/orders/{id}/label": {
      "get": {
        "operationId": "AdminOrdersController_label",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Seller-order Elchi QR yorlig‘ini PDF olish",
        "tags": [
          "admin-orders"
        ]
      }
    },
    "/api/v1/admin/orders/labels": {
      "post": {
        "operationId": "AdminOrdersController_labelsBatch",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShippingLabelsBatchDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Bir nechta seller-order yorlig‘ini bitta PDF olish",
        "tags": [
          "admin-orders"
        ]
      }
    },
    "/api/v1/admin/orders/{id}/cancel": {
      "post": {
        "operationId": "AdminOrdersController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminOrderActionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-orders"
        ]
      }
    },
    "/api/v1/admin/orders/{id}/refund": {
      "post": {
        "operationId": "AdminOrdersController_refund",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminOrderRefundDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-orders"
        ]
      }
    },
...
  },
  "info": {
    "title": "Elchi Marketplace API",
    "description": "Marketplace API — MVP (batafsil: docs/API_CONTRACT.md)",
    "version": "0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ApiErrorDetailDto": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "example": ""
          },
          "error": {
            "type": "string",
            "example": "phone +998XXXXXXXXX formatida bo‘lishi kerak"
          }
        },
        "required": [
          "field",
          "error"
        ]
      },
      "AuthErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "message": {
            "oneOf": [
              {
                "type": "string",
                "example": "Telefon yoki parol xato"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "example": [
                  "password must be longer than or equal to 8 characters"
                ]
              }
            ]
          },
          "errorCode": {
            "type": "string",
            "example": "VALIDATION_ERROR"
          },
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiErrorDetailDto"
            }
          }
        },
        "required": [
          "statusCode",
          "message",
          "errorCode"
        ]
      },
...
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "ShippingLabelsBatchDto": {
        "type": "object",
        "properties": {
          "orderIds": {
            "example": [
              "101",
              "102"
            ],
            "minItems": 1,
            "maxItems": 100,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "orderIds"
        ]
      },
...
      "AdminOrderActionDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "example": "Mahsulot mavjud emas"
          }
        },
        "required": [
          "reason"
        ]
      },
      "AdminOrderRefundDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "example": "Mahsulot mavjud emas"
          },
          "amount": {
            "type": "number",
            "example": 125000,
            "description": "Hozir faqat to‘liq refund: summa order totaliga teng bo‘lishi kerak"
          }
        },
        "required": [
          "reason"
        ]
      },
...
    }
  }
}